Update targetSdk to 35 and use enableEdgeToEdge
Resolves: #9429 #9459
