+++ feature/launcher/src/main/kotlin/app/k9mail/feature/launcher/FeatureLauncherActivity.kt
@@ -4,7 +4,6 @@ import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
 import app.k9mail.feature.launcher.ui.FeatureLauncherApp
 import com.fsck.k9.ui.base.K9Activity
 
@@ -12,9 +11,6 @@ class FeatureLauncherActivity : K9Activity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
-        enableEdgeToEdge()
-
         setContent {
             FeatureLauncherApp()
         }
